navigate searchinput
