+++ app/src/main/java/com/wafflestudio/waffleseminar2024/viewPagerFragments/SearchResultFragment.kt
@@ -1,0 +1,17 @@
+package com.wafflestudio.waffleseminar2024.viewPagerFragments
+
+import android.os.Bundle
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import androidx.fragment.app.Fragment
+import com.wafflestudio.waffleseminar2024.R
+
+class SearchResultFragment : Fragment() {
+    override fun onCreateView(
+        inflater: LayoutInflater, container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View? {
+        return inflater.inflate(R.layout.page_app, container, false)
+    }
+}
