+++ app/src/main/res/navigation/search_nav_graph.xml
@@ -1,0 +1,37 @@
+<?xml version="1.0" encoding="utf-8"?>
+<navigation xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    app:startDestination="@id/searchOverviewFragment">
+
+    <fragment
+        android:id="@+id/searchOverviewFragment"
+        android:name="com.wafflestudio.waffleseminar2024.viewPagerFragments.SearchOverviewFragment"
+        android:label="Search Overview">
+        <action
+            android:id="@+id/action_to_searchInputFragment"
+            app:destination="@id/searchInputFragment" />
+    </fragment>
+
+    <fragment
+        android:id="@+id/searchInputFragment"
+        android:name="com.wafflestudio.waffleseminar2024.viewPagerFragments.SearchInputFragment"
+        android:label="Search Input">
+        <action
+            android:id="@+id/action_to_searchResultFragment"
+            app:destination="@id/searchResultFragment" />
+    </fragment>
+
+    <fragment
+        android:id="@+id/searchResultFragment"
+        android:name="com.wafflestudio.waffleseminar2024.viewPagerFragments.SearchResultFragment"
+        android:label="Search Result">
+        <action
+            android:id="@+id/action_to_movieDetailFragment"
+            app:destination="@id/movieDetailFragment" />
+    </fragment>
+
+    <fragment
+        android:id="@+id/movieDetailFragment"
+        android:name="com.wafflestudio.waffleseminar2024.viewPagerFragments.MovieDetailFragment"
+        android:label="Movie Detail" />
+</navigation>
